Revert "Fleet UI: Fix host details > software vuln filter z index issue"
This reverts commit 598fdf021c2552308adcfd370f39f37bae76536f.
