+++ frontend/pages/hosts/details/HostDetailsPage/HostDetailsPage.tsx
@@ -39,13 +39,7 @@ import {
   CERTIFICATES_DEFAULT_SORT,
 } from "interfaces/certificates";
 
-import {
-  getNextLocationPath,
-  normalizeEmptyValues,
-  wrapFleetHelper,
-} from "utilities/helpers";
-import { convertParamsToSnakeCase } from "utilities/url";
-
+import { normalizeEmptyValues, wrapFleetHelper } from "utilities/helpers";
 import permissions from "utilities/permissions";
 import {
   DOCUMENT_TITLE_SUFFIX,
@@ -56,12 +50,6 @@ import {
 } from "utilities/constants";
 
 import { isAndroid, isIPadOrIPhone, isLinuxLike } from "interfaces/platform";
-import {
-  buildSoftwareFilterQueryParams,
-  buildSoftwareVulnFiltersQueryParams,
-  getSoftwareVulnFiltersFromQueryParams,
-  ISoftwareVulnFiltersParams,
-} from "pages/SoftwarePage/SoftwareTitles/SoftwareTable/helpers";
 
 import Spinner from "components/Spinner";
 import TabNav from "components/TabNav";
@@ -78,7 +66,6 @@ import {
   IPackageInstallDetails,
 } from "components/ActivityDetails/InstallDetails/SoftwareInstallDetails/SoftwareInstallDetails";
 import SoftwareUninstallDetailsModal from "components/ActivityDetails/InstallDetails/SoftwareUninstallDetailsModal";
-import SoftwareFiltersModal from "pages/SoftwarePage/components/SoftwareFiltersModal";
 import { IShowActivityDetailsData } from "components/ActivityItem/ActivityItem";
 
 import HostSummaryCard from "../cards/HostSummary";
@@ -126,7 +113,6 @@ import {
   generateUsernameValues,
 } from "../cards/User/helpers";
 import HostHeader from "../cards/HostHeader";
-import { getHostSoftwareFilterFromQueryParams } from "../cards/Software/helpers";
 
 const baseClass = "host-details";
 
@@ -143,7 +129,6 @@ interface IHostDetailsProps {
       query?: string;
       order_key?: string;
       order_direction?: "asc" | "desc";
-      per_page?: string;
     };
     search?: string;
   };
@@ -205,9 +190,6 @@ const HostDetailsPage = ({
   const [showLockHostModal, setShowLockHostModal] = useState(false);
   const [showUnlockHostModal, setShowUnlockHostModal] = useState(false);
   const [showWipeModal, setShowWipeModal] = useState(false);
-  const [showSoftwareFiltersModal, setShowSoftwareFiltersModal] = useState(
-    false
-  );
 
   // Used in activities to show run script details modal
   const [scriptExecutionId, setScriptExecutiontId] = useState("");
@@ -603,69 +585,6 @@ const HostDetailsPage = ({
     setShowUnenrollMdmModal(!showUnenrollMdmModal);
   }, [showUnenrollMdmModal, setShowUnenrollMdmModal]);
 
-  const toggleSoftwareFiltersModal = useCallback(() => {
-    setShowSoftwareFiltersModal(!showSoftwareFiltersModal);
-  }, [setShowSoftwareFiltersModal, showSoftwareFiltersModal]);
-
-  /**  Compares vuln filters to current vuln query params */
-  const determineVulnFilterChange = useCallback(
-    (vulnFilters: ISoftwareVulnFiltersParams) => {
-      const queryParams = parseHostSoftwareQueryParams(location.query);
-      const changedEntry = Object.entries(vulnFilters).find(([key, val]) => {
-        switch (key) {
-          case "vulnerable":
-          case "exploit": {
-            // Normalize values: undefined → false, then compare
-            const current = queryParams[key] ?? false;
-            const incoming = val ?? false;
-            return incoming !== current;
-          }
-          case "minCvssScore":
-            return val !== queryParams.min_cvss_score;
-          case "maxCvssScore":
-            return val !== queryParams.max_cvss_score;
-          default:
-            return false;
-        }
-      });
-      return changedEntry?.[0] ?? "";
-    },
-    [location.query]
-  );
-
-  const onApplyVulnFilters = (vulnFilters: ISoftwareVulnFiltersParams) => {
-    const queryParams = location.query;
-
-    const newQueryParams = {
-      query: queryParams.query,
-      orderDirection: queryParams.order_direction,
-      orderKey: queryParams.order_key,
-      perPage: queryParams.per_page,
-      page: 0, // resets page index
-      ...buildSoftwareFilterQueryParams(
-        getHostSoftwareFilterFromQueryParams(queryParams)
-      ),
-      ...buildSoftwareVulnFiltersQueryParams(vulnFilters),
-    };
-
-    // We want to determine which query param has changed in order to
-    // reset the page index to 0 if any other param has changed.
-    const changedParam = determineVulnFilterChange(vulnFilters);
-
-    // Update the route only if a change is detected
-    if (changedParam) {
-      router.replace(
-        getNextLocationPath({
-          pathPrefix: location.pathname,
-          routeTemplate: "",
-          queryParams: convertParamsToSnakeCase(newQueryParams),
-        })
-      );
-    }
-
-    toggleSoftwareFiltersModal();
-  };
-
   const onDestroyHost = async () => {
     if (host) {
       setIsUpdatingHost(true);
@@ -982,363 +901,358 @@ const HostDetailsPage = ({
 
   return (
     <MainContent className={baseClass}>
-      <HostDetailsBanners
-        mdmEnrollmentStatus={host?.mdm.enrollment_status}
-        hostPlatform={host?.platform}
-        macDiskEncryptionStatus={host?.mdm.macos_settings?.disk_encryption}
-        connectedToFleetMdm={host?.mdm.connected_to_fleet}
-        diskEncryptionOSSetting={host?.mdm.os_settings?.disk_encryption}
-        diskIsEncrypted={host?.disk_encryption_enabled}
-        diskEncryptionKeyAvailable={host?.mdm.encryption_key_available}
-      />
-      <div className={`${baseClass}__header-links`}>
-        <BackLink
-          text="Back to all hosts"
-          path={filteredHostsPath || PATHS.MANAGE_HOSTS}
+      <>
+        <HostDetailsBanners
+          mdmEnrollmentStatus={host?.mdm.enrollment_status}
+          hostPlatform={host?.platform}
+          macDiskEncryptionStatus={host?.mdm.macos_settings?.disk_encryption}
+          connectedToFleetMdm={host?.mdm.connected_to_fleet}
+          diskEncryptionOSSetting={host?.mdm.os_settings?.disk_encryption}
+          diskIsEncrypted={host?.disk_encryption_enabled}
+          diskEncryptionKeyAvailable={host?.mdm.encryption_key_available}
         />
-      </div>
-      <div className={`${baseClass}__header-summary`}>
-        <HostHeader
-          summaryData={summaryData}
-          showRefetchSpinner={showRefetchSpinner}
-          onRefetchHost={onRefetchHost}
-          renderActionDropdown={renderActionDropdown}
-          hostMdmDeviceStatus={hostMdmDeviceStatus}
-        />
-      </div>
-      <TabNav className={`${baseClass}__tab-nav`}>
-        <Tabs
-          selectedIndex={getTabIndex(location.pathname)}
-          onSelect={(i) => navigateToNav(i)}
-        >
-          <TabList>
-            {hostDetailsSubNav.map((navItem) => {
-              // Bolding text when the tab is active causes a layout shift
-              // so we add a hidden pseudo element with the same text string
-              return (
-                <Tab key={navItem.title}>
-                  <TabText count={navItem.count} isErrorCount>
-                    {navItem.name}
-                  </TabText>
-                </Tab>
-              );
-            })}
-          </TabList>
-          <TabPanel className={`${baseClass}__details-panel`}>
-            <HostSummaryCard
-              summaryData={summaryData}
-              bootstrapPackageData={bootstrapPackageData}
-              isPremiumTier={isPremiumTier}
-              toggleOSSettingsModal={toggleOSSettingsModal}
-              toggleBootstrapPackageModal={toggleBootstrapPackageModal}
-              hostSettings={host?.mdm.profiles ?? []}
-              osSettings={host?.mdm.os_settings}
-              osVersionRequirement={getOSVersionRequirementFromMDMConfig(
-                host.platform
-              )}
-              className={fullWidthCardClass}
-            />
-            <AboutCard
-              className={showUsersCard ? defaultCardClass : fullWidthCardClass}
-              aboutData={aboutData}
-              munki={macadmins?.munki}
-              mdm={mdm}
-            />
-            {showUsersCard && (
-              <UserCard
-                className={defaultCardClass}
-                platform={host.platform}
-                endUsers={host.end_users ?? []}
-                enableAddEndUser={
-                  isDarwinHost &&
-                  generateUsernameValues(host.end_users ?? []).length === 0
-                }
-                onAddEndUser={() => setShowAddEndUserModal(true)}
+        <div className={`${baseClass}__header-links`}>
+          <BackLink
+            text="Back to all hosts"
+            path={filteredHostsPath || PATHS.MANAGE_HOSTS}
+          />
+        </div>
+        <div className={`${baseClass}__header-summary`}>
+          <HostHeader
+            summaryData={summaryData}
+            showRefetchSpinner={showRefetchSpinner}
+            onRefetchHost={onRefetchHost}
+            renderActionDropdown={renderActionDropdown}
+            hostMdmDeviceStatus={hostMdmDeviceStatus}
+          />
+        </div>
+        <TabNav className={`${baseClass}__tab-nav`}>
+          <Tabs
+            selectedIndex={getTabIndex(location.pathname)}
+            onSelect={(i) => navigateToNav(i)}
+          >
+            <TabList>
+              {hostDetailsSubNav.map((navItem) => {
+                // Bolding text when the tab is active causes a layout shift
+                // so we add a hidden pseudo element with the same text string
+                return (
+                  <Tab key={navItem.title}>
+                    <TabText count={navItem.count} isErrorCount>
+                      {navItem.name}
+                    </TabText>
+                  </Tab>
+                );
+              })}
+            </TabList>
+            <TabPanel className={`${baseClass}__details-panel`}>
+              <HostSummaryCard
+                summaryData={summaryData}
+                bootstrapPackageData={bootstrapPackageData}
+                isPremiumTier={isPremiumTier}
+                toggleOSSettingsModal={toggleOSSettingsModal}
+                toggleBootstrapPackageModal={toggleBootstrapPackageModal}
+                hostSettings={host?.mdm.profiles ?? []}
+                osSettings={host?.mdm.os_settings}
+                osVersionRequirement={getOSVersionRequirementFromMDMConfig(
+                  host.platform
+                )}
+                className={fullWidthCardClass}
               />
-            )}
-            {showActivityCard && (
-              <ActivityCard
+              <AboutCard
                 className={
-                  showAgentOptionsCard
-                    ? doubleHeightCardClass
+                  showUsersCard ? defaultCardClass : fullWidthCardClass
+                }
+                aboutData={aboutData}
+                munki={macadmins?.munki}
+                mdm={mdm}
+              />
+              {showUsersCard && (
+                <UserCard
+                  className={defaultCardClass}
+                  platform={host.platform}
+                  endUsers={host.end_users ?? []}
+                  enableAddEndUser={
+                    isDarwinHost &&
+                    generateUsernameValues(host.end_users ?? []).length === 0
+                  }
+                  onAddEndUser={() => setShowAddEndUserModal(true)}
+                />
+              )}
+              {showActivityCard && (
+                <ActivityCard
+                  className={
+                    showAgentOptionsCard
+                      ? doubleHeightCardClass
+                      : defaultCardClass
+                  }
+                  activeTab={activeActivityTab}
+                  activities={
+                    activeActivityTab === "past"
+                      ? pastActivities
+                      : upcomingActivities
+                  }
+                  isLoading={
+                    activeActivityTab === "past"
+                      ? pastActivitiesIsFetching
+                      : upcomingActivitiesIsFetching
+                  }
+                  isError={
+                    activeActivityTab === "past"
+                      ? pastActivitiesIsError
+                      : upcomingActivitiesIsError
+                  }
+                  canCancelActivities={
+                    isGlobalAdmin ||
+                    isGlobalMaintainer ||
+                    isHostTeamAdmin ||
+                    isHostTeamMaintainer
+                  }
+                  upcomingCount={upcomingActivities?.count || 0}
+                  onChangeTab={onChangeActivityTab}
+                  onNextPage={() => setActivityPage(activityPage + 1)}
+                  onPreviousPage={() => setActivityPage(activityPage - 1)}
+                  onShowDetails={onShowActivityDetails}
+                  onCancel={onCancelActivity}
+                />
+              )}
+              {showAgentOptionsCard && (
+                <AgentOptionsCard
+                  className={defaultCardClass}
+                  osqueryData={osqueryData}
+                  wrapFleetHelper={wrapFleetHelper}
+                  isChromeOS={host?.platform === "chrome"}
+                />
+              )}
+              <LabelsCard
+                className={
+                  !showActivityCard && !showAgentOptionsCard
+                    ? fullWidthCardClass
                     : defaultCardClass
                 }
-                activeTab={activeActivityTab}
-                activities={
-                  activeActivityTab === "past"
-                    ? pastActivities
-                    : upcomingActivities
-                }
-                isLoading={
-                  activeActivityTab === "past"
-                    ? pastActivitiesIsFetching
-                    : upcomingActivitiesIsFetching
-                }
-                isError={
-                  activeActivityTab === "past"
-                    ? pastActivitiesIsError
-                    : upcomingActivitiesIsError
-                }
-                canCancelActivities={
-                  isGlobalAdmin ||
-                  isGlobalMaintainer ||
-                  isHostTeamAdmin ||
-                  isHostTeamMaintainer
-                }
-                upcomingCount={upcomingActivities?.count || 0}
-                onChangeTab={onChangeActivityTab}
-                onNextPage={() => setActivityPage(activityPage + 1)}
-                onPreviousPage={() => setActivityPage(activityPage - 1)}
-                onShowDetails={onShowActivityDetails}
-                onCancel={onCancelActivity}
+                labels={host?.labels || []}
+                onLabelClick={onLabelClick}
               />
-            )}
-            {showAgentOptionsCard && (
-              <AgentOptionsCard
-                className={defaultCardClass}
-                osqueryData={osqueryData}
-                wrapFleetHelper={wrapFleetHelper}
-                isChromeOS={host?.platform === "chrome"}
+              {showLocalUserAccountsCard && (
+                <LocalUserAccountsCard
+                  className={fullWidthCardClass}
+                  users={host?.users || []}
+                  usersState={usersState}
+                  isLoading={isLoadingHost}
+                  onUsersTableSearchChange={onUsersTableSearchChange}
+                  hostUsersEnabled={featuresConfig?.enable_host_users}
+                />
+              )}
+              {showCertificatesCard && (
+                <CertificatesCard
+                  className={fullWidthCardClass}
+                  data={hostCertificates}
+                  hostPlatform={host.platform}
+                  onSelectCertificate={onSelectCertificate}
+                  isError={isErrorHostCertificates}
+                  page={certificatePage}
+                  pageSize={DEFAULT_CERTIFICATES_PAGE_SIZE}
+                  onNextPage={() => setCertificatePage(certificatePage + 1)}
+                  onPreviousPage={() => setCertificatePage(certificatePage - 1)}
+                  sortDirection={sortCerts.order_direction}
+                  sortHeader={sortCerts.order_key}
+                  onSortChange={setSortCerts}
+                />
+              )}
+            </TabPanel>
+            <TabPanel>
+              <SoftwareCard
+                id={host.id}
+                platform={host.platform}
+                softwareUpdatedAt={host.software_updated_at}
+                hostCanWriteSoftware={!!host.orbit_version || isIosOrIpadosHost}
+                hostScriptsEnabled={host.scripts_enabled || false}
+                isSoftwareEnabled={featuresConfig?.enable_software_inventory}
+                router={router}
+                queryParams={parseHostSoftwareQueryParams(location.query)}
+                pathname={location.pathname}
+                onShowSoftwareDetails={setSelectedSoftwareDetails}
+                hostTeamId={host.team_id || 0}
+                hostMDMEnrolled={host.mdm.connected_to_fleet}
               />
-            )}
-            <LabelsCard
-              className={
-                !showActivityCard && !showAgentOptionsCard
-                  ? fullWidthCardClass
-                  : defaultCardClass
-              }
-              labels={host?.labels || []}
-              onLabelClick={onLabelClick}
-            />
-            {showLocalUserAccountsCard && (
-              <LocalUserAccountsCard
-                className={fullWidthCardClass}
-                users={host?.users || []}
-                usersState={usersState}
-                isLoading={isLoadingHost}
-                onUsersTableSearchChange={onUsersTableSearchChange}
-                hostUsersEnabled={featuresConfig?.enable_host_users}
-              />
-            )}
-            {showCertificatesCard && (
-              <CertificatesCard
-                className={fullWidthCardClass}
-                data={hostCertificates}
+              {isDarwinHost && macadmins?.munki?.version && (
+                <MunkiIssuesCard
+                  isLoading={isLoadingHost}
+                  munkiIssues={macadmins.munki_issues}
+                  deviceType={host?.platform === "darwin" ? "macos" : ""}
+                />
+              )}
+            </TabPanel>
+            <TabPanel>
+              <QueriesCard
+                hostId={host.id}
+                router={router}
                 hostPlatform={host.platform}
-                onSelectCertificate={onSelectCertificate}
-                isError={isErrorHostCertificates}
-                page={certificatePage}
-                pageSize={DEFAULT_CERTIFICATES_PAGE_SIZE}
-                onNextPage={() => setCertificatePage(certificatePage + 1)}
-                onPreviousPage={() => setCertificatePage(certificatePage - 1)}
-                sortDirection={sortCerts.order_direction}
-                sortHeader={sortCerts.order_key}
-                onSortChange={setSortCerts}
+                schedule={schedule}
+                queryReportsDisabled={
+                  config?.server_settings?.query_reports_disabled
+                }
               />
-            )}
-          </TabPanel>
-          <TabPanel>
-            <SoftwareCard
-              id={host.id}
-              platform={host.platform}
-              softwareUpdatedAt={host.software_updated_at}
-              hostCanWriteSoftware={!!host.orbit_version || isIosOrIpadosHost}
-              hostScriptsEnabled={host.scripts_enabled || false}
-              isSoftwareEnabled={featuresConfig?.enable_software_inventory}
-              router={router}
-              queryParams={parseHostSoftwareQueryParams(location.query)}
-              pathname={location.pathname}
-              onShowSoftwareDetails={setSelectedSoftwareDetails}
-              hostTeamId={host.team_id || 0}
-              hostMDMEnrolled={host.mdm.connected_to_fleet}
-              toggleSoftwareFiltersModal={toggleSoftwareFiltersModal}
-            />
-            {isDarwinHost && macadmins?.munki?.version && (
-              <MunkiIssuesCard
+              {canViewPacks && (
+                <PacksCard packsState={packsState} isLoading={isLoadingHost} />
+              )}
+            </TabPanel>
+            <TabPanel>
+              <PoliciesCard
+                policies={host?.policies || []}
                 isLoading={isLoadingHost}
-                munkiIssues={macadmins.munki_issues}
-                deviceType={host?.platform === "darwin" ? "macos" : ""}
+                togglePolicyDetailsModal={togglePolicyDetailsModal}
+                hostPlatform={host.platform}
+                router={router}
+                currentTeamId={currentTeam?.id}
               />
-            )}
-          </TabPanel>
-          <TabPanel>
-            <QueriesCard
-              hostId={host.id}
-              router={router}
-              hostPlatform={host.platform}
-              schedule={schedule}
-              queryReportsDisabled={
-                config?.server_settings?.query_reports_disabled
-              }
-            />
-            {canViewPacks && (
-              <PacksCard packsState={packsState} isLoading={isLoadingHost} />
-            )}
-          </TabPanel>
-          <TabPanel>
-            <PoliciesCard
-              policies={host?.policies || []}
-              isLoading={isLoadingHost}
-              togglePolicyDetailsModal={togglePolicyDetailsModal}
-              hostPlatform={host.platform}
-              router={router}
-              currentTeamId={currentTeam?.id}
-            />
-          </TabPanel>
-        </Tabs>
-      </TabNav>
-      {showDeleteHostModal && (
-        <DeleteHostModal
-          onCancel={() => setShowDeleteHostModal(false)}
-          onSubmit={onDestroyHost}
-          hostName={host?.display_name}
-          isUpdating={isUpdatingHost}
-        />
-      )}
-      {showSelectQueryModal && host && (
-        <SelectQueryModal
-          onCancel={() => setShowSelectQueryModal(false)}
-          isOnlyObserver={isOnlyObserver}
-          hostId={hostIdFromURL}
-          hostTeamId={host?.team_id}
-          router={router}
-          currentTeamId={currentTeam?.id}
-        />
-      )}
-      {showScriptModalGroup && (
-        <ScriptModalGroup
-          host={host}
-          currentUser={currentUser}
-          onCloseScriptModalGroup={onCloseScriptModalGroup}
-          teamIdForApi={currentTeam?.id}
-        />
-      )}
-      {!!host && showTransferHostModal && (
-        <TransferHostModal
-          onCancel={() => setShowTransferHostModal(false)}
-          onSubmit={onTransferHostSubmit}
-          teams={teams || []}
-          isGlobalAdmin={isGlobalAdmin as boolean}
-          isUpdating={isUpdatingHost}
-        />
-      )}
-      {!!host && showPolicyDetailsModal && (
-        <PolicyDetailsModal
-          onCancel={onCancelPolicyDetailsModal}
-          policy={selectedPolicy}
-        />
-      )}
-      {showOSSettingsModal && (
-        <OSSettingsModal
-          canResendProfiles={host.platform === "darwin"}
-          hostId={host.id}
-          platform={host.platform}
-          hostMDMData={host.mdm}
-          onClose={toggleOSSettingsModal}
-          onProfileResent={refetchHostDetails}
-        />
-      )}
-      {showUnenrollMdmModal && !!host && (
-        <UnenrollMdmModal hostId={host.id} onClose={toggleUnenrollMdmModal} />
-      )}
-      {showDiskEncryptionModal && host && (
-        <DiskEncryptionKeyModal
-          platform={host.platform}
-          hostId={host.id}
-          onCancel={() => setShowDiskEncryptionModal(false)}
-        />
-      )}
-      {showBootstrapPackageModal &&
-        bootstrapPackageData.details &&
-        bootstrapPackageData.name && (
-          <BootstrapPackageModal
-            packageName={bootstrapPackageData.name}
-            details={bootstrapPackageData.details}
-            onClose={() => setShowBootstrapPackageModal(false)}
+            </TabPanel>
+          </Tabs>
+        </TabNav>
+        {showDeleteHostModal && (
+          <DeleteHostModal
+            onCancel={() => setShowDeleteHostModal(false)}
+            onSubmit={onDestroyHost}
+            hostName={host?.display_name}
+            isUpdating={isUpdatingHost}
           />
         )}
-      {scriptExecutionId && (
-        <RunScriptDetailsModal
-          scriptExecutionId={scriptExecutionId}
-          onCancel={onCancelRunScriptDetailsModal}
-        />
-      )}
-      {!!packageInstallDetails && (
-        <SoftwareInstallDetailsModal
-          details={packageInstallDetails}
-          onCancel={onCancelSoftwareInstallDetailsModal}
-        />
-      )}
-      {packageUninstallDetails && (
-        <SoftwareUninstallDetailsModal
-          details={packageUninstallDetails}
-          onCancel={() => setPackageUninstallDetails(null)}
-        />
-      )}
-      {!!appInstallDetails && (
-        <AppInstallDetailsModal
-          details={appInstallDetails}
-          onCancel={onCancelAppInstallDetailsModal}
-        />
-      )}
-      {showLockHostModal && (
-        <LockModal
-          id={host.id}
-          platform={host.platform}
-          hostName={host.display_name}
-          onSuccess={() => setHostMdmDeviceState("locking")}
-          onClose={() => setShowLockHostModal(false)}
-        />
-      )}
-      {showUnlockHostModal && (
-        <UnlockModal
-          id={host.id}
-          platform={host.platform}
-          hostName={host.display_name}
-          onSuccess={() => {
-            host.platform !== "darwin" && setHostMdmDeviceState("unlocking");
-          }}
-          onClose={() => setShowUnlockHostModal(false)}
-        />
-      )}
-      {showWipeModal && (
-        <WipeModal
-          id={host.id}
-          hostName={host.display_name}
-          onSuccess={() => setHostMdmDeviceState("wiping")}
-          onClose={() => setShowWipeModal(false)}
-        />
-      )}
-      {selectedSoftwareDetails && (
-        <SoftwareDetailsModal
-          hostDisplayName={host.display_name}
-          software={selectedSoftwareDetails}
-          onExit={() => setSelectedSoftwareDetails(null)}
-        />
-      )}
-      {selectedCancelActivity && (
-        <CancelActivityModal
-          hostId={host.id}
-          activity={selectedCancelActivity}
-          onCancelActivity={() => refetchUpcomingActivities()}
-          onSuccessCancel={onSuccessCancelActivity}
-          onExit={() => setSelectedCancelActivity(null)}
-        />
-      )}
-      {selectedCertificate && (
-        <CertificateDetailsModal
-          certificate={selectedCertificate}
-          onExit={() => setSelectedCertificate(null)}
-        />
-      )}
-      {showSoftwareFiltersModal && (
-        <SoftwareFiltersModal
-          onExit={toggleSoftwareFiltersModal}
-          onSubmit={onApplyVulnFilters}
-          vulnFilters={getSoftwareVulnFiltersFromQueryParams(location.query)}
-          isPremiumTier={isPremiumTier || false}
-        />
-      )}
+        {showSelectQueryModal && host && (
+          <SelectQueryModal
+            onCancel={() => setShowSelectQueryModal(false)}
+            isOnlyObserver={isOnlyObserver}
+            hostId={hostIdFromURL}
+            hostTeamId={host?.team_id}
+            router={router}
+            currentTeamId={currentTeam?.id}
+          />
+        )}
+        {showScriptModalGroup && (
+          <ScriptModalGroup
+            host={host}
+            currentUser={currentUser}
+            onCloseScriptModalGroup={onCloseScriptModalGroup}
+            teamIdForApi={currentTeam?.id}
+          />
+        )}
+        {!!host && showTransferHostModal && (
+          <TransferHostModal
+            onCancel={() => setShowTransferHostModal(false)}
+            onSubmit={onTransferHostSubmit}
+            teams={teams || []}
+            isGlobalAdmin={isGlobalAdmin as boolean}
+            isUpdating={isUpdatingHost}
+          />
+        )}
+        {!!host && showPolicyDetailsModal && (
+          <PolicyDetailsModal
+            onCancel={onCancelPolicyDetailsModal}
+            policy={selectedPolicy}
+          />
+        )}
+        {showOSSettingsModal && (
+          <OSSettingsModal
+            canResendProfiles={host.platform === "darwin"}
+            hostId={host.id}
+            platform={host.platform}
+            hostMDMData={host.mdm}
+            onClose={toggleOSSettingsModal}
+            onProfileResent={refetchHostDetails}
+          />
+        )}
+        {showUnenrollMdmModal && !!host && (
+          <UnenrollMdmModal hostId={host.id} onClose={toggleUnenrollMdmModal} />
+        )}
+        {showDiskEncryptionModal && host && (
+          <DiskEncryptionKeyModal
+            platform={host.platform}
+            hostId={host.id}
+            onCancel={() => setShowDiskEncryptionModal(false)}
+          />
+        )}
+        {showBootstrapPackageModal &&
+          bootstrapPackageData.details &&
+          bootstrapPackageData.name && (
+            <BootstrapPackageModal
+              packageName={bootstrapPackageData.name}
+              details={bootstrapPackageData.details}
+              onClose={() => setShowBootstrapPackageModal(false)}
+            />
+          )}
+        {scriptExecutionId && (
+          <RunScriptDetailsModal
+            scriptExecutionId={scriptExecutionId}
+            onCancel={onCancelRunScriptDetailsModal}
+          />
+        )}
+        {!!packageInstallDetails && (
+          <SoftwareInstallDetailsModal
+            details={packageInstallDetails}
+            onCancel={onCancelSoftwareInstallDetailsModal}
+          />
+        )}
+        {packageUninstallDetails && (
+          <SoftwareUninstallDetailsModal
+            details={packageUninstallDetails}
+            onCancel={() => setPackageUninstallDetails(null)}
+          />
+        )}
+        {!!appInstallDetails && (
+          <AppInstallDetailsModal
+            details={appInstallDetails}
+            onCancel={onCancelAppInstallDetailsModal}
+          />
+        )}
+        {showLockHostModal && (
+          <LockModal
+            id={host.id}
+            platform={host.platform}
+            hostName={host.display_name}
+            onSuccess={() => setHostMdmDeviceState("locking")}
+            onClose={() => setShowLockHostModal(false)}
+          />
+        )}
+        {showUnlockHostModal && (
+          <UnlockModal
+            id={host.id}
+            platform={host.platform}
+            hostName={host.display_name}
+            onSuccess={() => {
+              host.platform !== "darwin" && setHostMdmDeviceState("unlocking");
+            }}
+            onClose={() => setShowUnlockHostModal(false)}
+          />
+        )}
+        {showWipeModal && (
+          <WipeModal
+            id={host.id}
+            hostName={host.display_name}
+            onSuccess={() => setHostMdmDeviceState("wiping")}
+            onClose={() => setShowWipeModal(false)}
+          />
+        )}
+        {selectedSoftwareDetails && (
+          <SoftwareDetailsModal
+            hostDisplayName={host.display_name}
+            software={selectedSoftwareDetails}
+            onExit={() => setSelectedSoftwareDetails(null)}
+          />
+        )}
+        {selectedCancelActivity && (
+          <CancelActivityModal
+            hostId={host.id}
+            activity={selectedCancelActivity}
+            onCancelActivity={() => refetchUpcomingActivities()}
+            onSuccessCancel={onSuccessCancelActivity}
+            onExit={() => setSelectedCancelActivity(null)}
+          />
+        )}
+        {selectedCertificate && (
+          <CertificateDetailsModal
+            certificate={selectedCertificate}
+            onExit={() => setSelectedCertificate(null)}
+          />
+        )}
+      </>
       {showAddEndUserModal && (
         <AddEndUserModal onExit={() => setShowAddEndUserModal(false)} />
       )}
